refactor: useGame 훅 리팩토링
- useMultiGame 생성
- useGame은 useSoloGame과 useMultiGame 내부에서 사용하는 훅으로 변경
- 멀티게임과 솔로게임 페이지 분리 준비
