+++ src/app/game/multi/[gameId]/page.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import Players from '@/components/game/players/Players';
+import PlayZone from '@/components/game/playZone/PlayZone';
+
+const MultiGame = () => {
+  return (
+    <>
+      <Players />
+      <PlayZone />
+    </>
+  );
+};
+
+export default MultiGame;
