+++ src/app/game/solo/[gameId]/page.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const SoloGame = () => {
+  return (
+    <section>
+      <div>SoloGame</div>
+    </section>
+  );
+};
+
+export default SoloGame;
